WEBREF-110 :truck: Renamed calendar icon to date
